Reformat for line length

+++ dataworkspace/dataworkspace/apps/applications/admin.py
@@ -71,10 +71,7 @@ class ApplicationInstanceAdmin(admin.ModelAdmin):
 
         return '{0:.2f}% at {1}'.format(
             max_cpu,
-            datetime.datetime.fromtimestamp(ts_at_max)
-            .strftime('%-I:%M %p')
-            .replace('AM', 'a.m.')
-            .replace('PM', 'p.m'),
+            datetime.datetime.fromtimestamp(ts_at_max).strftime('%-I:%M %p').replace('AM', 'a.m.').replace('PM', 'p.m'),
         )
 
     max_cpu.short_description = 'Max recent CPU'
@@ -100,16 +97,11 @@ class ApplicationFilter(SimpleListFilter):
     template = 'admin/application_instance_report_filter.html'
 
     def lookups(self, request, model_admin):
-        return tuple(
-            (app.id, app.nice_name)
-            for app in ApplicationTemplate.objects.all().order_by('nice_name')
-        )
+        return tuple((app.id, app.nice_name) for app in ApplicationTemplate.objects.all().order_by('nice_name'))
 
     def queryset(self, request, queryset):
         try:
-            queryset_filter = {
-                'application_template__id': request.GET['application_template__id']
-            }
+            queryset_filter = {'application_template__id': request.GET['application_template__id']}
         except KeyError:
             queryset_filter = {}
         return queryset.filter(**queryset_filter)
@@ -134,11 +126,8 @@ class ApplicationGroup(SimpleListFilter):
     def choices(self, changelist):
         for i, (lookup, title) in enumerate(self.lookup_choices):
             yield {
-                'selected': self.value() == str(lookup)
-                or (i == 0 and self.value() is None),
-                'query_string': changelist.get_query_string(
-                    {self.parameter_name: lookup}
-                ),
+                'selected': self.value() == str(lookup) or (i == 0 and self.value() is None),
+                'query_string': changelist.get_query_string({self.parameter_name: lookup}),
                 'display': title,
             }
 
@@ -179,12 +168,8 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
             # NULL values are ordered as greater than non-NULL values, so to order rows without
             # runtime as lower in the list as those that have runtime, but still order rows with
             # runtime in decreasing order, we need an extra field
-            'has_runtime': Least(
-                Count(F('spawner_stopped_at') - F('spawner_created_at')), 1
-            ),
-            'num_with_runtime': Count(
-                F('spawner_stopped_at') - F('spawner_created_at')
-            ),
+            'has_runtime': Least(Count(F('spawner_stopped_at') - F('spawner_created_at')), 1),
+            'num_with_runtime': Count(F('spawner_stopped_at') - F('spawner_created_at')),
             'min_runtime': Min(
                 Func(Value('second'), F('spawner_stopped_at'), function='date_trunc')
                 - Func(Value('second'), F('spawner_created_at'), function='date_trunc')
@@ -200,10 +185,7 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
         }
 
         group_by_fields = {
-            'user_and_application': [
-                'owner__username',
-                'application_template__nice_name',
-            ],
+            'user_and_application': ['owner__username', 'application_template__nice_name',],
             'user': ['owner__username'],
             'application': ['application_template__nice_name'],
             'user_and_cpu_memory': ['owner__username', 'spawner_cpu', 'spawner_memory'],
@@ -213,21 +195,12 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
         summary_with_applications = list(
             qs.values(*group_by_fields)
             .annotate(**metrics)
-            .order_by(
-                *(
-                    ['-has_runtime', '-total_runtime', '-num_launched', '-max_runtime']
-                    + group_by_fields
-                )
-            )
+            .order_by(*(['-has_runtime', '-total_runtime', '-num_launched', '-max_runtime'] + group_by_fields))
         )
 
         perm = list(Permission.objects.filter(codename='start_all_applications'))
         users = (
-            User.objects.filter(
-                Q(groups__permissions__in=perm)
-                | Q(user_permissions__in=perm)
-                | Q(is_superuser=True)
-            )
+            User.objects.filter(Q(groups__permissions__in=perm) | Q(user_permissions__in=perm) | Q(is_superuser=True))
             .distinct()
             .order_by('username')
         )
@@ -245,9 +218,7 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
         )
 
         def group_by_user_missing_rows():
-            users_with_applications = set(
-                item['owner__username'] for item in summary_with_applications
-            )
+            users_with_applications = set(item['owner__username'] for item in summary_with_applications)
             return [
                 {
                     'owner__username': user.username,
@@ -261,13 +232,8 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
             ]
 
         def group_by_application_missing_rows():
-            application_templates = list(
-                ApplicationTemplate.objects.filter(**app_filter).order_by('nice_name')
-            )
-            applications_run = set(
-                item['application_template__nice_name']
-                for item in summary_with_applications
-            )
+            application_templates = list(ApplicationTemplate.objects.filter(**app_filter).order_by('nice_name'))
+            applications_run = set(item['application_template__nice_name'] for item in summary_with_applications)
             return [
                 {
                     'owner__username': None,
@@ -282,8 +248,7 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
 
         def group_by_cpu_memory_missing_rows():
             launched_cpu_memory_combos = set(
-                (item['spawner_cpu'], item['spawner_memory'])
-                for item in summary_with_applications
+                (item['spawner_cpu'], item['spawner_memory']) for item in summary_with_applications
             )
             return [
                 {
@@ -316,12 +281,9 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
             ]
 
         def group_by_user_and_application_missing_rows():
-            application_templates = list(
-                ApplicationTemplate.objects.filter(**app_filter).order_by('nice_name')
-            )
+            application_templates = list(ApplicationTemplate.objects.filter(**app_filter).order_by('nice_name'))
             users_with_applications = set(
-                (item['owner__username'], item['application_template__nice_name'])
-                for item in summary_with_applications
+                (item['owner__username'], item['application_template__nice_name']) for item in summary_with_applications
             )
             return [
                 {
@@ -332,8 +294,7 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
                     'num_with_runtime': 0,
                 }
                 for user, application_template in product(users, application_templates)
-                if (user.username, application_template.nice_name)
-                not in users_with_applications
+                if (user.username, application_template.nice_name) not in users_with_applications
             ]
 
         summary_without_applications = (
@@ -348,9 +309,7 @@ class ApplicationInstanceReportAdmin(admin.ModelAdmin):
             else group_by_user_and_application_missing_rows()
         )
 
-        response.context_data['summary'] = (
-            summary_with_applications + summary_without_applications
-        )
+        response.context_data['summary'] = summary_with_applications + summary_without_applications
 
         response.context_data['summary_total'] = dict(qs.aggregate(**metrics))
 
@@ -361,9 +320,7 @@ class VisualisationTemplateEditForm(forms.ModelForm):
     authorized_master_datasets = forms.ModelMultipleChoiceField(
         required=False,
         widget=FilteredSelectMultiple('master datasets', False),
-        queryset=MasterDataset.objects.live()
-        .filter(user_access_type='REQUIRES_AUTHORIZATION')
-        .order_by('name'),
+        queryset=MasterDataset.objects.live().filter(user_access_type='REQUIRES_AUTHORIZATION').order_by('name'),
     )
 
     class Meta:
@@ -378,9 +335,7 @@ class VisualisationTemplateEditForm(forms.ModelForm):
         except KeyError:
             return
 
-        self.fields[
-            'authorized_master_datasets'
-        ].initial = MasterDataset.objects.live().filter(
+        self.fields['authorized_master_datasets'].initial = MasterDataset.objects.live().filter(
             datasetapplicationtemplatepermission__application_template=instance
         )
 
@@ -430,23 +385,15 @@ class VisualisationTemplateAdmin(admin.ModelAdmin):
             )
 
         current_master_datasets = set(
-            DataSet.objects.live().filter(
-                datasetapplicationtemplatepermission__application_template=obj
-            )
+            DataSet.objects.live().filter(datasetapplicationtemplatepermission__application_template=obj)
         )
-        authorized_master_datasets = set(
-            form.cleaned_data['authorized_master_datasets']
-        )
+        authorized_master_datasets = set(form.cleaned_data['authorized_master_datasets'])
 
         for dataset in authorized_master_datasets - current_master_datasets:
-            DataSetApplicationTemplatePermission.objects.create(
-                dataset=dataset, application_template=obj
-            )
+            DataSetApplicationTemplatePermission.objects.create(dataset=dataset, application_template=obj)
             log_change('Added dataset {} permission'.format(dataset))
         for dataset in current_master_datasets - authorized_master_datasets:
-            DataSetApplicationTemplatePermission.objects.filter(
-                dataset=dataset, application_template=obj
-            ).delete()
+            DataSetApplicationTemplatePermission.objects.filter(dataset=dataset, application_template=obj).delete()
             log_change('Removed dataset {} permission'.format(dataset))
 
         super().save_model(request, obj, form, change)
